+++ src/input/input_source_interpolation.cc
@@ -75,7 +75,7 @@ InputSourceInterpolation::InputSourceInterpolation(
   }
 
   remained_num_samples_ = frame_period_ + 1;
-  data_length_ = curr_data_.size();
+  data_length_ = static_cast<int>(curr_data_.size());
 
   if (!source_->Get(&next_data_)) {
     next_data_.resize(data_length_);
